% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_installation.R
\name{check_package_installation}
\alias{check_package_installation}
\title{Check package installation}
\usage{
check_package_installation(pkg, fun = NULL, task = NULL)
}
\arguments{
\item{pkg}{(\emph{character}) the name of the suggested R package}

\item{fun}{(\emph{character}) the name of the function that uses \code{pkg}}

\item{task}{(\emph{character}; \emph{optional}) the task that \code{fun} completes.}
}
\value{
\code{TRUE}, invisibly, if \code{pkg} is installed.
}
\description{
Several secondary R packages are used and it's preferable to
install them only if they are needed. Placing this check in the body of a
function will provide helpful instructions for cases when a user attempts
to run a function but doesn't have the required suggested package(s).
(The main inspiration of this was because a lot of people had issues installing
some packages needed for only specific analysis)
}
\examples{
\dontrun{
result <- MotrpacHumanPreSuspensionAnalysis::check_package_installation(pkg = 'base', fun = 'mean')
result
}
}
