% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_TISSUE_ABBR}
\alias{HUMAN_TISSUE_ABBR}
\title{Tissue abbreviations for compact figure annotation}
\format{
A named character vector of tissue abbreviations.
}
\usage{
HUMAN_TISSUE_ABBR
}
\description{
Named character vector providing short tissue abbreviations for use in
heatmap annotations, labels, and compact visual encodings.
}
\keyword{datasets}
