% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OME_TISSUE_CODE.R
\docType{data}
\name{OME_TISSUE_CODE}
\alias{OME_TISSUE_CODE}
\title{Tissue codes relevant for a given ome}
\format{
A \code{data.frame} object with a tissue, the relevant BIC tissue
code, and the ome for the given tissue.
}
\usage{
OME_TISSUE_CODE
}
\description{
Data frame of proper tissue codes for each ome/tissue, which
is especially important to get a physiological understanding of different
types of tissues (e.g. whole blood vs pbmcs)
}
\keyword{datasets}
