% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_enrichment.R
\name{plot_cluster_enrichment}
\alias{plot_cluster_enrichment}
\title{Bubble heatmap of CAMERA-PR or ORA results from fuzzy c-means
clustering}
\usage{
plot_cluster_enrichment(
  x,
  set_ids = NULL,
  selected_tissues = c("adipose", "blood", "muscle"),
  selected_ome = c("transcript-rna-seq", "prot-pr", "prot-ph", "metab"),
  padj_cutoff = 0.05,
  n_top = 6L,
  column_title = NULL,
  filename
)
}
\arguments{
\item{x}{a \code{data.frame}. Either
\code{\link[MotrpacHumanPreSuspensionAnalysis]{FCM_CAMERA}} or
\code{\link[MotrpacHumanPreSuspensionAnalysis]{FCM_ORA}}.}

\item{set_ids}{character or \code{NULL}; one or more of \code{x$set_id}. The
dataset \code{x} will be filtered to these sets. If \code{NULL}, the
\code{n_top} most significant sets from each contrast will be selected.}

\item{selected_tissues}{character; one or more tissues that will be used to
create the heatmap. If multiple tissues are selected, only those terms
tested in at least two tissues will be plotted.}

\item{selected_ome}{character; the ome that will be used to create the
heatmap.}

\item{padj_cutoff}{numeric; cutoff in terms of p-value for an asterisk to be included}

\item{n_top}{integer; if \code{set_ids} is \code{NULL}, the \code{n_top} most
significant molecular signatures from each tissue and cluster will be
selected.}

\item{column_title}{character or \code{NULL}; title for the heatmap.}

\item{filename}{character; path to a PDF file to save the heatmap.}
}
\value{
Nothing. The heatmap is saved to a PDF specified by \code{filename}.
}
\description{
Generate a bubble heatmap from the results of nonparametric
CAMERA-PR or ORA applied to the output of fuzzy c-means clustering.
}
\author{
Tyler Sagendorf
}
