% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmeans.R
\name{plot_cmeans}
\alias{plot_cmeans}
\title{Save a Plot of Fuzzy C-Means Cluster Trajectories}
\usage{
plot_cmeans(
  FCM,
  filename,
  keep_clusters,
  min_membership = 0,
  common_ylim = TRUE,
  ncol = 4L,
  dpi = 250,
  modality = c("both", "Endur", "Resist")
)
}
\arguments{
\item{FCM}{an object of class \code{fclust}: one tissue's element of
\code{\link[MotrpacHumanPreSuspensionAnalysis]{FCM_CLUSTERS}} or of the
list returned by \code{\link{run_cmeans}}. See examples.}

\item{filename}{File name to create on disk.}

\item{keep_clusters}{integer; the identifier numbers of clusters to include
in the plot. Defaults to all clusters.}

\item{min_membership}{numeric; to be included in the plot, the probability
that a feature belongs to a cluster should be \eqn{\geq}
\code{min_membership}.}

\item{common_ylim}{logical; whether the plots for each cluster should have
the same y-axis limits.}

\item{ncol}{integer; the maximum number of clusters to place on a single row.
The number of rows is determined automatically.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Only applies when converting pixel units,
as is typical for raster output types.}

\item{modality}{character; which exercise modalities should be used for
plotting? One of \code{"both"}, \code{"Endur"}, or \code{"Resist"}.}
}
\description{
Essentially, a ggplot2 version of \code{Mfuzz::plot.mfuzz2} with
a different color scheme that saves a plot to a file, rather than
displaying it.
}
\examples{
\dontrun{
# Pre-computed FCM results
plot_cmeans(FCM = FCM_CLUSTERS[["adipose"]],
            filename = "adipose_clusters.pdf",
            min_membership = 0.3)

# New FCM results
FCM <- run_cmeans(selected_tissues = "adipose", modality = "Endur")

plot_cmeans(FCM = FCM[["adipose"]],
            filename = "adipose_endur_clusters.pdf",
            modality = "Endur")
}
}
\seealso{
\code{\link{run_cmeans}},
\code{\link[MotrpacHumanPreSuspensionAnalysis]{FCM_CLUSTERS}}
}
\author{
Tyler Sagendorf
}
