% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precovid_cca.R
\name{plot_precovid_cca}
\alias{plot_precovid_cca}
\title{canCor analysis cor correlation to principal components}
\usage{
plot_precovid_cca(
  data_to_pca,
  sample_metadata,
  to_scale = TRUE,
  custom_title = "",
  interesting_variables = c("pid", "randomGroupCode", "Timepoint", "Sex", "BMI",
    "calculatedAge"),
  num_pcs = 5
)
}
\arguments{
\item{data_to_pca}{The actual numerical data that should be included in analysis.
must be able to be converted into a numerical matrix}

\item{sample_metadata}{The correponding metadata to \code{data_to_pca}. The
rownames of this matrix must correspond to colnames of the data matrix.}

\item{to_scale}{logical; Whether to scale the data when performing PCA.}

\item{custom_title}{character; whatever title is desired for the figure}

\item{interesting_variables}{character; a vector of variables that are
going to be correlated using canCors to the PCs. The # of PCs in
\code{num_pcs} are included by default.}

\item{num_pcs}{numeric; the number of principal components desired to be included.}
}
\value{
a pheatmap object with the desired correlations to principal components
}
\description{
Performs PCA and correlates PCs to selected metadata using
\code{variancePartition::canCorPairs}.
}
\author{
Christopher Jin
}
