% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precovid_pca.R
\name{plot_precovid_pca}
\alias{plot_precovid_pca}
\title{Principal component analysis for the pre-cawg group}
\usage{
plot_precovid_pca(
  data_to_pca,
  sample_metadata,
  to_scale = TRUE,
  custom_title = ""
)
}
\arguments{
\item{data_to_pca}{The actual numerical data that should be included in analysis.
must be able to be converted into a numerical matrix}

\item{sample_metadata}{The correponding metadata to \code{data_to_pca}. The
rownames of this matrix must correspond to colnames of the data matrix.}

\item{to_scale}{logical; Whether to scale the data when performing PCA.}

\item{custom_title}{character; whatever title is desired for the figure}
}
\value{
a ggplot object with the desired PCA plotting aesthetic.
}
\description{
Basically just a wrapper function to plot with shapes and colors
that fits the pre-cawg figure guidelines#'
}
