% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ORA.R
\name{run_ORA}
\alias{run_ORA}
\title{Over-Representation Analysis (ORA)}
\usage{
run_ORA(
  input,
  background,
  database = names(MotrpacHumanPreSuspensionAnalysis::MOLECULAR_SIGNATURES),
  path_to_gmt = NULL,
  min_size = 5L,
  overlap_cutoff = 0.7
)
}
\arguments{
\item{input}{character; vector of "interesting" features. Most likely genes,
but may be RefMet metabolite IDs or singly-phosphorylated peptides.}

\item{background}{character; vector of all features that were analyzed. Used
to filter the molecular signatures.}

\item{database}{character; one or more names specifying the database(s) to
test. Options are (case insensitive) \code{"BIOCARTA"},
\code{"KEGG_MEDICUS"}, \code{"PID"}, \code{"REACTOME"}, \code{"WP"}
(WikiPathways database), \code{"GOBP"}, \code{"GOCC"}, \code{"GOMF"},
\code{"MITOCARTA"} (MitoCarta3.0 database), \code{"PSP"} (PhosphoSitePlus
kinases; only valid when \code{selected_omes} contains \code{"prot-ph"}),
or \code{"REFMET"} (RefMet chemical subclasses; only valid when
\code{selected_omes} contains \code{"metab"}). See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{MOLECULAR_SIGNATURES}} for
details.}

\item{path_to_gmt}{character; (optional) path to one or more GMT files.
Passed to \code{TMSig::readGMT}. If provided, \code{database} is ignored.}

\item{min_size}{integer; the minimum set size for testing.}

\item{overlap_cutoff}{numeric; the minimum proportion of genes in each set
that must appear in a given dataset. Used to pre-filter sets. Does not
affect \code{"metab"} or \code{"prot-ph"} results. This will always be 0.1
for \code{"prot-ol"} results.}
}
\value{
An object of class \code{data.frame} with the following columns:
\describe{
\item{\code{collection}}{factor; the broad molecular signature
collection. Only included when \code{path_to_gmt} is \code{NULL}. See
\code{\link{SET_TO_ID}} for details.}
\item{\code{database}}{factor; the molecular signature database. Only
included when \code{path_to_gmt} is \code{NULL}. See
\code{\link{SET_TO_ID}} for details.}
\item{\code{set_id}}{character; a unique ID for the molecular
signature. Only included when \code{path_to_gmt} is \code{NULL}. See
\code{\link{SET_TO_ID}} for details.}
\item{\code{set}}{character; the molecular signature being tested.
For global proteomics and transcriptomics, these are gene sets. For
phosphoproteomics, these are kinase sets.}
\item{\code{set_short}}{character; a shortened version of
\code{set}. Only included when \code{path_to_gmt} is \code{NULL}. See
\code{\link{SET_TO_ID}} for details.}
\item{\code{set_size}}{integer; the number of molecules in the set that
were present in the \code{background} vector.}
\item{\code{set_size_DB}}{integer; the number of molecules in the set,
as defined in \code{\link{MOLECULAR_SIGNATURES}}.}
\item{\code{size_ratio}}{numeric; the ratio of \code{set_size} to
\code{set_size_DB}, rounded to the nearest thousandth. A measure of
confidence that the gene set being tested is correctly described by the
entry in the \code{set} column. While smaller values do not
necessarily indicate that the results are unreliable, terms from the gene
set databases should be treated with caution.}
\item{\code{set_size_in_input}}{integer; the number of molecules in the
set that were present in the \code{input} vector.}
\item{\code{input_size}}{integer; the size of the \code{input} vector.}
\item{\code{background_size}}{integer; the size of the \code{background}
vector.}
\item{\code{p_value}}{numeric; the two-sided p-value.}
\item{\code{adj_p_value}}{numeric; the BH-adjusted p-value. P-values
are adjusted within each combination of tissue, assay, contrast, and
collection.}
}
}
\description{
Fast over-representation analysis.
}
\examples{
# Use genes from all gene sets as the background
bg <- MotrpacHumanPreSuspensionAnalysis::MOLECULAR_SIGNATURES
bg[c("PSP", "REFMET")] <- NULL
bg <- unique(unlist(bg))

# Use all genes from the MitoCarta OXPHOS term as the interesting subset
input <- MotrpacHumanPreSuspensionAnalysis::MOLECULAR_SIGNATURES$MITOCARTA[["MITOCARTA_OXPHOS"]]

res <- run_ORA(input = input, background = bg)
head(res)
}
\seealso{
\code{\link{run_cluster_ORA}}, \code{\link{run_cameraPR}}
}
