% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCION.R
\name{run_SCION}
\alias{run_SCION}
\title{@title SCION implementation for preCAWG.
Combines preprocessing and network inference into one function}
\usage{
run_SCION(
  randomGroupCode = c("ADUResist", "ADUEndur"),
  regulators,
  targets,
  permute = NULL,
  dim = "col",
  cluster = TRUE,
  dir.name = "exp",
  weightthreshold = 0,
  normalize = FALSE,
  num.cores = 1,
  connect.hubs = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{randomGroupCode}{\itemize{
\item regulators, targets, and clustering is all done on a
per-modality basis. All the clustering is done using z-scores from the differential analysis
}}

\item{regulators}{the set of hypothesized tissue/omes that regulate the
targets matrix, from .load_scion_matrixes}

\item{targets}{the set of hypothesized tissue/omes that are targetted by the
regulators matrix, from \code{.load_scion_matrixes}}

\item{permute}{number of random permutations to perform for edge trimming.
Default NULL, which means no permutations will be performed}

\item{dim}{dimension on which to permute. options are "row" or "col". default "col"}

\item{cluster}{boolean option to cluster using c-means prior to network inference. default TRUE}

\item{dir.name}{name of directory to save results. default "exp".
if using permutations, this parameter is ignored, and the directories are
named after the permutation number (1,2,3...)}

\item{weightthreshold}{threshold for edge trimming. all edges with weight < weightthreshold
are removed from the network. minimum value of 0. default 0.}

\item{normalize}{boolean to normalize edge weights to a 0-1 scale. default FALSE}

\item{num.cores}{number of cores to use for parallelization. num.cores-1 will
be used for parallelization. if num.cores < 3, no parallelization is performed. default 1.}

\item{connect.hubs}{boolean to connect the hubs between clusters.
this parameter is ignored if clustering is not performed. default TRUE}

\item{verbose}{boolean to display detailed output. default FALSE}
}
\description{
This function is based on the source code \code{https://github.com/nmclark2/SCION}
}
\author{
Natalie M Clark, Christopher Jin
}
