% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SCION.R
\name{scion_run_cmeans}
\alias{scion_run_cmeans}
\title{Clustering for the scion function}
\usage{
scion_run_cmeans(DA_Input_Cmeans, num_clusters = 13)
}
\arguments{
\item{DA_Input_Cmeans}{list of differential results that fits the cmeans function}

\item{num_clusters}{numeric; number of desired clusters.}
}
\value{
A list with relevant clustering result information
}
\description{
Scales data and runs clustering to find groups to form networks
from for the scion function
}
