% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ome_tissue.R
\name{tissue_available_list}
\alias{tissue_available_list}
\title{List the Available Tissues}
\usage{
tissue_available_list(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; whether to output messages.}
}
\value{
A character vector of the available tissues: "adipose", "blood", and
"muscle".
}
\description{
Vector of tissues available in the analysis. These are the "full
names" for each of the tissues.
}
\examples{
tissue_available_list()
}
