% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_RESULTS.R
\docType{data}
\name{CLIN_CHEMISTRY_DA}
\alias{CLIN_CHEMISTRY_DA}
\title{Differential Alternative Splicing Results}
\format{
A data frame containing all differential analysis results for clinical metabolites.
}
\source{
Generated as described in \code{data-raw/generate_differential_analysis}
}
\usage{
CLIN_CHEMISTRY_DA
}
\description{
A data object containing results from a differential analysis for the clinical chemistry analytes.
}
\keyword{datasets}
