% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_RESULTS.R
\docType{data}
\name{CLIN_CHEMISTRY_DA}
\alias{CLIN_CHEMISTRY_DA}
\title{Clinical Chemistry Differential analysis}
\format{
A \code{data.frame} with 297 rows and 20 columns.
}
\usage{
CLIN_CHEMISTRY_DA
}
\description{
A data object containing results from a differential expression analysis of the clinical
analytes as described in data-raw/differential_analysis_results/
}
\keyword{datasets}
