% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_RESULTS.R
\docType{data}
\name{DA_RESULTS}
\alias{DA_RESULTS}
\alias{ADIPOSE_METAB_DA}
\alias{ADIPOSE_PROT_PH_DA}
\alias{ADIPOSE_PROT_PR_DA}
\alias{ADIPOSE_TRNSCRPT_DA}
\alias{BLOOD_METAB_DA}
\alias{BLOOD_PROT_OL_DA}
\alias{BLOOD_TRNSCRPT_DA}
\alias{MUSCLE_METAB_DA}
\alias{MUSCLE_PROT_PH_DA}
\alias{MUSCLE_PROT_PR_DA}
\alias{MUSCLE_TRNSCRPT_DA}
\alias{BLOOD_METAB_T_CLINICAL_DA}
\alias{BLOOD_PROT_CLINICAL_DA}
\title{Differential Analysis Results}
\format{
These objects are not to be used directly. See
\code{\link{load_differential_analysis}} for format details.
}
\usage{
ADIPOSE_METAB_DA
ADIPOSE_PROT_PH_DA
ADIPOSE_PROT_PR_DA
ADIPOSE_TRNSCRPT_DA

BLOOD_METAB_DA
BLOOD_PROT_OL_DA
BLOOD_METAB_T_CLINICAL_DA
BLOOD_PROT_CLINICAL_DA
BLOOD_TRNSCRPT_DA

MUSCLE_METAB_DA
MUSCLE_PROT_PH_DA
MUSCLE_PROT_PR_DA
MUSCLE_TRNSCRPT_DA
}
\description{
Differential analysis results. Not to be used directly! Instead,
use \code{\link{load_differential_analysis}}.
}
\keyword{datasets}
