% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ome_tissue.R
\name{metab_only_list}
\alias{metab_only_list}
\title{List the Metabolomics Platforms}
\usage{
metab_only_list()
}
\value{
A character vector containing the targeted and untargeted
metabolomics platforms.
}
\description{
A character vector containing the targeted and untargeted
metabolomics platforms.
}
\examples{
metab_only_list()
}
