% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METABOLOMICS_CVS.R
\docType{data}
\name{METABOLOMICS_CVS}
\alias{METABOLOMICS_CVS}
\title{Metabolite Coefficients of Variation}
\format{
A \code{data.frame} object with description of tissue, platform, and
whether or not a given platform is lowest CV.
}
\usage{
METABOLOMICS_CVS
}
\description{
Metabolite coefficients of variation. Each coefficient of variation per platform was measured on reference standards that were processed through each platform. For more information, see the QC Notebook in the precovid-analyses repository
}
\keyword{datasets}
