% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ome_tissue.R
\name{ome_available_list}
\alias{ome_available_list}
\title{List All Omes}
\usage{
ome_available_list()
}
\value{
A character vector containing all available omes.
}
\description{
Return a character vector containing all available omes. These
are the "full names" for each of the assays.
}
\examples{
ome_available_list()
}
