% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUTLIERS.R
\docType{data}
\name{OUTLIERS}
\alias{OUTLIERS}
\title{List of outliers designated by the Pre-CAWG working group.}
\format{
An object of class \code{data.frame} with 154 rows and 4 columns.
}
\usage{
OUTLIERS
}
\description{
Data object combining outlier forms from each omic-analyst.
The process for flagging a sample is described in more detail in the methods,
but, generally, samples that were >3x IQR in the first 3 principal
components were flagged and manually reviewed. Differs by ome.
}
\keyword{datasets}
