% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cameraPR.R
\name{run_cameraPR}
\alias{run_cameraPR}
\title{Wrapper for CAMERA-PR}
\usage{
run_cameraPR(
  DA_list = NULL,
  selected_omes = c("transcript-rna-seq", "prot-pr", "prot-ph", "prot-ol", "metab"),
  selected_tissues = "all",
  database = setdiff(names(MotrpacHumanPreSuspensionAnalysis::MOLECULAR_SIGNATURES),
    "PTMSIGDB"),
  path_to_gmt = NULL,
  min_size = 5L,
  overlap_cutoff = 0.7
)
}
\arguments{
\item{DA_list}{list; a named list of \code{data.frame} objects, each
containing differential analysis results for a specific tissue/assay
combination. The list must be nested, with tissues at the top level and
omes within tissues. If \code{NULL} (default), the differential analysis
results will be generated with
\code{\link[MotrpacHumanPreSuspensionAnalysis]{load_differential_analysis}}.
Unless wishing to analyze DA results that are not in
MotrpacHumanPreSuspensionAnalysis, this should remain \code{NULL}.}

\item{selected_omes}{character; one or more character strings selected from
the following options: \code{"transcript-rna-seq"}, \code{"prot-pr"},
\code{"prot-ph"}, and \code{"metab"} (all metabolomics platforms). Passed
to \code{\link[MotrpacHumanPreSuspensionAnalysis]{load_differential_analysis}}.}

\item{selected_tissues}{character; passed to
\code{\link[MotrpacHumanPreSuspensionAnalysis]{load_differential_analysis}}.
One or more of the following: \code{"all"}, \code{"muscle"},
\code{"adipose"}, or \code{"blood"}.}

\item{database}{character; one or more names specifying the database(s) to
test. Options are (case insensitive) \code{"BIOCARTA"},
\code{"KEGG_MEDICUS"}, \code{"PID"}, \code{"REACTOME"}, \code{"WP"}
(WikiPathways database), \code{"GOBP"}, \code{"GOCC"}, \code{"GOMF"},
\code{"MITOCARTA"} (MitoCarta3.0 database), \code{"PSP"} (PhosphoSitePlus
kinases; only valid when \code{selected_omes} contains \code{"prot-ph"}),
or \code{"REFMET"} (RefMet chemical subclasses; only valid when
\code{selected_omes} contains \code{"metab"}). See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{MOLECULAR_SIGNATURES}} for
details.}

\item{path_to_gmt}{character; (optional) path to one or more GMT files.
Passed to \code{TMSig::readGMT}. If provided, \code{database} is ignored.}

\item{min_size}{integer; the minimum set size for testing.}

\item{overlap_cutoff}{numeric; the minimum proportion of genes in each set
that must appear in a given dataset. Used to pre-filter sets. Does not
affect \code{"metab"} or \code{"prot-ph"} results. This will always be 0.1
for \code{"prot-ol"} results.}
}
\value{
An object of class \code{data.frame} with the following columns:
\describe{
\item{\code{tissue}}{factor; the tissue.}

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\item\{\code{assay}\}\{factor; the omics assay.\}

\\item\{\code{contrast_type}\}\{factor; the type of contrast.\}

\\item\{\code{contrast}\}\{factor; the contrast of interest.\}

\\item\{\code{contrast_short}\}\{factor; shortened contrasts.\}

\\item\{\code{collection}\}\{factor; the broad molecular signature
collection. Only included when \code{path_to_gmt} is \code{NULL}. See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.\}

\\item\{\code{database}\}\{factor; the molecular signature database. Only
included when \code{path_to_gmt} is \code{NULL}. See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.\}

\\item\{\code{set_id}\}\{character; a unique ID for the molecular
signature. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}}
for details.\}

\\item\{\code{set}\}\{character; the molecular signature being tested.
For global proteomics and transcriptomics, these are gene sets. For
phosphoproteomics, these are kinase sets.\}

\\item\{\code{set_short}\}\{character; a shortened version of
\code{set}. Only included when \code{path_to_gmt} is \code{NULL}. See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.\}

\\item\{\code{set_size}\}\{integer; the number of molecules in the set that
were present in the DA results for that specific tissue/assay
combination.\}

\\item\{\code{set_size_DB}\}\{integer; the number of molecules in the set,
as defined in the GMT file.\}

\\item\{\code{size_ratio}\}\{numeric; the ratio of \code{set_size} to
\code{set_size_DB}, rounded to the nearest thousandth. A measure of
confidence that the gene set being tested is correctly described by the
entry in the \code{set} column. While smaller values do not
necessarily indicate that the results are unreliable, terms from the gene
set databases should be treated with caution.\}

\\item\{\code{direction}\}\{factor; the direction of change: \code{"Up"}
or \code{"Down"}.\}

\\item\{\code{t}\}\{numeric; the two-sample t-statistic.\}

\\item\{\code{df}\}\{integer; the available degrees of freedom.\}

\\item\{\code{z.std}\}\{numeric; standard Normal equivalents of \code{t}.\}

\\item\{\code{p_value}\}\{numeric; the two-sided p-value.\}

\\item\{\code{adj_p_value}\}\{numeric; the BH-adjusted p-value. P-values
are adjusted within each combination of tissue, assay, contrast, and
collection.\}
}\if{html}{\out{</div>}}

}
}
\description{
A wrapper for \code{\link[TMSig]{cameraPR.matrix}} that performs
pre-ranked Correlation Adjusted MEan RAnk molecular signature analysis of
differential analysis results.
}
\examples{
\dontrun{
# Test Reactome and Gene Ontology Biological Processes
# databases on muscle global proteomics.
x1 <- run_cameraPR(selected_omes = "prot-pr",
                   selected_tissues = "muscle",
                   database = c("GOBP", "REACTOME"))

# Test all omes using all molecular signature databases
x2 <- run_cameraPR()
}
}
\seealso{
\code{\link[MotrpacHumanPreSuspensionAnalysis]{MOLECULAR_SIGNATURES}},
\code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}}
}
\author{
Tyler Sagendorf
}
