% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cluster_ORA.R
\name{run_cluster_ORA}
\alias{run_cluster_ORA}
\title{Analyze Fuzzy C-means Clustering Results with ORA}
\usage{
run_cluster_ORA(
  FCM,
  selected_omes = c("transcript-rna-seq", "prot-pr", "prot-ph", "metab"),
  selected_tissues = "all",
  database = names(MotrpacHumanPreSuspensionAnalysis::MOLECULAR_SIGNATURES),
  path_to_gmt = NULL,
  min_size = 5L,
  overlap_cutoff = 0.7,
  min_prob = 0.3
)
}
\arguments{
\item{FCM}{fuzzy c-means (FCM) clustering results. Output of
\code{\link{run_cmeans}}. This should be a named list of objects of class
\code{fclust}.}

\item{selected_omes}{character; one or more character strings selected from
the following options: \code{"transcript-rna-seq"}, \code{"prot-pr"},
\code{"prot-ph"}, and \code{"metab"} (all metabolomics platforms). Passed
to \code{\link[MotrpacHumanPreSuspensionAnalysis]{load_differential_analysis}}.}

\item{selected_tissues}{character; passed to
\code{\link[MotrpacHumanPreSuspensionAnalysis]{load_differential_analysis}}.
One or more of the following: \code{"all"}, \code{"muscle"},
\code{"adipose"}, or \code{"blood"}.}

\item{database}{character; one or more names specifying the database(s) to
test. Options are (case insensitive) \code{"BIOCARTA"},
\code{"KEGG_MEDICUS"}, \code{"PID"}, \code{"REACTOME"}, \code{"WP"}
(WikiPathways database), \code{"GOBP"}, \code{"GOCC"}, \code{"GOMF"},
\code{"MITOCARTA"} (MitoCarta3.0 database), \code{"PSP"} (PhosphoSitePlus
kinases; only valid when \code{selected_omes} contains \code{"prot-ph"}),
or \code{"REFMET"} (RefMet chemical subclasses; only valid when
\code{selected_omes} contains \code{"metab"}). See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{MOLECULAR_SIGNATURES}} for
details.}

\item{path_to_gmt}{character; (optional) path to one or more GMT files.
Passed to \code{TMSig::readGMT}. If provided, \code{database} is ignored.}

\item{min_size}{integer; the minimum set size for testing.}

\item{overlap_cutoff}{numeric; the minimum proportion of genes in each set
that must appear in a given dataset. Used to pre-filter sets. Does not
affect \code{"metab"} or \code{"prot-ph"} results. This will always be 0.1
for \code{"prot-ol"} results.}

\item{min_prob}{numeric; the minimum cluster membership probability required
for features to belong to a cluster. All features will be used for the
background for ORA, regardless of membership probability.}
}
\value{
An object of class \code{data.frame} with the following columns:
\describe{
\item{\code{tissue}}{factor; the tissue.}
\item{\code{assay}}{factor; the omics assay.}
\item{\code{cluster}}{factor; the cluster number. The clusters have the
same meaning across omics assays measured in the same tissue, but they
have no such relationship across tissues.}
\item{\code{collection}}{factor; the broad molecular signature
collection. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}}
for details.}
\item{\code{database}}{factor; the molecular signature database. See
\code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}} for details.}
\item{\code{set_id}}{character; a unique ID for the molecular
signature. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}}
for details.}
\item{\code{set}}{character; the molecular signature being tested.
For global proteomics and transcriptomics, these are gene sets. For
phosphoproteomics, these are kinase sets.}
\item{\code{set_short}}{character; a shortened version of
\code{set}. See \code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}}
for details.}
\item{\code{set_size}}{integer; the number of molecules in the set that
were present in the DA results for that specific tissue/assay
combination.}
\item{\code{set_size_DB}}{integer; the number of molecules in the set,
as defined in the GMT file.}
\item{\code{size_ratio}}{numeric; the ratio of \code{set_size} to
\code{set_size_DB}, rounded to the nearest thousandth. A measure of
confidence that the gene set being tested is correctly described by the
entry in the \code{set} column. While smaller values do not
necessarily indicate that the results are unreliable, terms from the gene
set databases should be treated with caution.}
\item{\code{set_size_in_cluster}}{integer; the number of elements in the
set that localized to the cluster.}
\item{\code{cluster_size}}{integer; number of features in each cluster
with membership probabilities of at least \code{min_prob}.}
\item{\code{background_size}}{integer; the number of elements in the
background. Includes all features that appeared in clustering results.}
\item{\code{p_value}}{numeric; the upper-tail p-value.}
\item{\code{adj_p_value}}{numeric; the BH-adjusted p-value. P-values are
adjusted within each combination of tissue, assay, collection, and
cluster.}
}
}
\description{
Over-representation analysis (ORA) applied to results of fuzzy
c-means (FCM) clustering.
}
\examples{
\dontrun{
  FCM <- run_cmeans()

  # Run ORA with all available molecular signatures
  cluster_res <- run_cluster_ORA(FCM = FCM)
  head(cluster_res)
}
}
\seealso{
\code{\link{run_cmeans}}, \code{\link{run_cluster_cameraPR}},
\code{\link[MotrpacHumanPreSuspensionAnalysis]{MOLECULAR_SIGNATURES}},
\code{\link[MotrpacHumanPreSuspensionAnalysis]{SET_TO_ID}}
}
\author{
Tyler Sagendorf
}
