% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cmeans.R
\name{run_cmeans}
\alias{run_cmeans}
\title{Fuzzy C-Means (FCM) Clustering}
\usage{
run_cmeans(
  selected_tissues = c("all", "adipose", "blood", "muscle"),
  selected_omes = c("transcript-rna-seq", "prot-pr", "prot-ol", "prot-ph", "metab"),
  num_clusters_adipose = 13L,
  num_clusters_blood = 13L,
  num_clusters_muscle = 12L,
  modality = c("both", "Endur", "Resist")
)
}
\arguments{
\item{selected_tissues}{character; passed to
\code{\link[MotrpacHumanPreSuspensionAnalysis]{load_differential_analysis}}.
One or more of the following: \code{"all"}, \code{"muscle"},
\code{"adipose"}, or \code{"blood"}.}

\item{selected_omes}{character; one or more of \code{"transcript-rna-seq"},
\code{"prot-pr"}, \code{"prot-ol"}, \code{"prot-ph"}, or \code{"metab"}.}

\item{num_clusters_adipose, num_clusters_blood, num_clusters_muscle}{integer;
the number of clusters desired for each tissue. Defaults to 13 for adipose,
13 for blood, and 12 for muscle. If more than one number is provided,
\code{Mfuzz::Dmin} will be used to generate a plot of the minimum centroid
distances, and the user will be asked to specify the optimal cluster number
in the console for each tissue.}

\item{modality}{character; which exercise modalities should be used for FCM?
One of \code{"both"}, \code{"Endur"}, or \code{"Resist"}.}
}
\value{
A named list of objects where names are tissues. Each object is of
class \code{"fclust"} with additional list component \code{"input"} for the
matrix of scaled z-scores used as input for FCM (both modalities included,
even when \code{modality != "both"}).
}
\description{
Fuzzy C-Means (FCM) Clustering
}
\examples{
\dontrun{
  x1 <- run_cmeans()
  names(x1) # list available components

  # Try a range of cluster numbers for one tissue
  x2 <- run_cmeans(selected_tissues = "adipose",
                   num_clusters_adipose = 3:14)

  # FCM for a single modality
  x3 <- run_cmeans(selected_tissues = "adipose",
                   modality = "Endur")
}
}
\seealso{
\code{\link{plot_cmeans}}, \code{\link{run_cluster_cameraPR}},
\code{\link{run_cluster_ORA}}
}
\author{
Tyler Sagendorf, Christopher Jin
}
